% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comptest-package.R
\docType{package}
\name{comptest-package}
\alias{comptest}
\alias{comptest-package}
\title{comptest: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tomletessier/comptest}
  \item Report bugs at \url{https://github.com/tomletessier/comptest/issues}
}

}
\author{
\strong{Maintainer}: Tom Bech Letessier \email{tomletessier@gmail.com} (\href{https://orcid.org/0000-0003-4011-0207}{ORCID}) [copyright holder]

}
\keyword{internal}
